created folder models, moved paddle to its own file in models/paddle.py
